Move FileReadMode to file_type_parser
