+++ airbyte-cdk/python/airbyte_cdk/sources/file_based/file_based_stream_reader.py
@@ -8,7 +8,8 @@ from io import IOBase
 from typing import Iterable, List, Optional, Set
 
 from airbyte_cdk.sources.file_based.config.abstract_file_based_spec import AbstractFileBasedSpec
-from airbyte_cdk.sources.file_based.remote_file import FileReadMode, RemoteFile
+from airbyte_cdk.sources.file_based.file_types.file_type_parser import FileReadMode
+from airbyte_cdk.sources.file_based.remote_file import RemoteFile
 from wcmatch.glob import GLOBSTAR, globmatch
