+++ airbyte-cdk/python/airbyte_cdk/sources/file_based/file_types/parquet_parser.py
@@ -10,8 +10,8 @@ import pyarrow as pa
 import pyarrow.parquet as pq
 from airbyte_cdk.sources.file_based.config.file_based_stream_config import FileBasedStreamConfig, ParquetFormat
 from airbyte_cdk.sources.file_based.file_based_stream_reader import AbstractFileBasedStreamReader
-from airbyte_cdk.sources.file_based.file_types.file_type_parser import FileTypeParser
-from airbyte_cdk.sources.file_based.remote_file import FileReadMode, RemoteFile
+from airbyte_cdk.sources.file_based.file_types.file_type_parser import FileReadMode, FileTypeParser
+from airbyte_cdk.sources.file_based.remote_file import RemoteFile
 from pyarrow import Scalar
