+++ airbyte-cdk/python/airbyte_cdk/sources/file_based/remote_file.py
@@ -3,15 +3,9 @@
 #
 
 from datetime import datetime
-from enum import Enum
 from typing import Optional
 
 from pydantic import BaseModel
-
-
-class FileReadMode(Enum):
-    READ = "r"
-    READ_BINARY = "rb"
 
 
 class RemoteFile(BaseModel):
