+++ airbyte-cdk/python/unit_tests/sources/file_based/helpers.py
@@ -11,8 +11,9 @@ from airbyte_cdk.sources.file_based.config.file_based_stream_config import FileB
 from airbyte_cdk.sources.file_based.discovery_policy import DefaultDiscoveryPolicy
 from airbyte_cdk.sources.file_based.file_based_stream_reader import AbstractFileBasedStreamReader
 from airbyte_cdk.sources.file_based.file_types.csv_parser import CsvParser
+from airbyte_cdk.sources.file_based.file_types.file_type_parser import FileReadMode
 from airbyte_cdk.sources.file_based.file_types.jsonl_parser import JsonlParser
-from airbyte_cdk.sources.file_based.remote_file import FileReadMode, RemoteFile
+from airbyte_cdk.sources.file_based.remote_file import RemoteFile
 from airbyte_cdk.sources.file_based.schema_validation_policies import AbstractSchemaValidationPolicy
 from unit_tests.sources.file_based.in_memory_files_source import InMemoryFilesStreamReader
